// src/modules/inventory/reports/finish-store/finish-fabric-receive-challan-report/finish-fabric-receive-challan-report-type.ts
export interface FinishFabricReceiveChallanReportType {
  COMPANY_NAME?: string;
  COMPANY_ADDRESS?: string;
  COMPANY_REMARKS?: string;
  BUYER_ID?: number;
  BUYER_NAME?: string;
  PO_ID?: number;
  PO_NO?: string;
  STYLE_ID?: number;
  STYLENAME?: string;
  STYLENO?: string;
  SUPPLIER_INTER_COMPANY_ID?: number;
  DYEING_FACTORY_ID?: number;
  DYEING_FACTORY?: string;
  SUPPLIER?: string;
  ID?: number;
  MASTER_ID?: number;
  FR_WORK_ORDER_ID?: number;
  GREY_SHAPE?: string;
  YARN_LOT_NO_ID?: number;
  YARN_LOT_NUMBER?: string;
  FABRIC_PART_ID?: number;
  USE_AREA?: string;
  GMT_COLOR_ID?: number;
  GMT_COLOR?: string;
  GMT_SIZE_ID?: number;
  GMT_SIZE?: string;
  SORTINGNO?: number;
  FIN_DIA?: number;
  FIN_REQ_SHAPE?: string;
  REMARKS?: string;
  ROLL_QTY?: number;
  GREY_WEIGHT?: number;
  QUANTITY?: number;
  PICES?: number;
  UOM_ID?: number;
  UOM?: string;
  FABRIC_TYPE_ID?: number;
  FABRIC_TYPE?: string;
  GSM?: number;
  FABRIC_ID?: number;
  FABRIC?: string;
  PROCESS_ID?: number;
  PROCESS_NAME?: string;
  OPERATION_SERIAL?: number;
  BATCH_NO?: string;
  COLOR_CODE?: string;
  WASTAGE?: number;
  YARN_CHALLAN_ID?: number;
  BB_LC_ID?: number;
  BBLC_NO?: string;
  WO_ID?: number;
  WORK_ORDER_NUMBER?: string;
  YARN_BRAND_ID?: number;
  YARN_BRAND?: string;
  STITCH_LENGTH?: number;
  GAUGE?: number;
  DYEING_BATCH_NO?: string;
  ISSUED_QTY?: number;
  CHECK_QTY?: number;
  SEASON_ID?: number;
  SEASON?: string;
  ROLL_NUMBER?: string;
  RATE?: number;
  PINO?: string;
}